chore: 사용자 인증 및 권한 검증하는 로직 분리
